09.1 -- ajout cors

+++ api/DTO/FavoriteDTO.cs
@@ -1,13 +1,15 @@
+using System;
+
 namespace DTO {
     public class FavoriteDTO
     {
         public FavoriteDTO()
         {
         }
-        
+
         public int Id {get; set;}
         public string UrlImage {get; set;}
         public string DressName {get; set;}
-        public double DressPrice {get; set;}
+        public Decimal DressPrice {get; set;}
     }
 }
